% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{in_range}
\alias{in_range}
\title{Test whether values are in an interval}
\usage{
in_range(x, lower, upper)
}
\arguments{
\item{x}{A numeric object.}

\item{lower, upper}{Numeric lower and upper bounds. Each must have size one
or a size compatible with `x`.}
}
\value{
A logical object with the shape of the longest argument.
}
\description{
Tests whether values fall in the closed interval defined by `lower` and
`upper`.
}
\details{
Missing values in `x` produce missing results. Bounds may be
  infinite, but must not be missing.
}
\examples{
in_range(1:5, 2, 4)

}
