% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{wrap}
\alias{wrap}
\title{Wrap values to a periodic interval}
\usage{
wrap(x, lower, upper)
}
\arguments{
\item{x}{A numeric object.}

\item{lower, upper}{Finite numeric bounds. Each must have size one or a size
compatible with `x`.}
}
\value{
A numeric object with the shape of the longest argument.
}
\description{
Wraps values to the half-open interval from `lower` to `upper`.
}
\details{
`lower` must be less than `upper`. Values equal to `upper` wrap to
  `lower`. Infinite values in `x` produce `NaN`.
}
\examples{
wrap(c(-10, 0, 370), 0, 360)

}
